requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
google-generativeai==0.8.6
python-dotenv==1.0.1
